Move getDisposalTextFromResult folder into lib/results folder
