+++ packages/core/lib/results/getDisposalTextFromResult/disqualifiedFromKeepingDisposalText.test.ts
