+++ packages/core/lib/results/getDisposalTextFromResult/disqualifiedFromKeepingDisposalText.ts
