+++ packages/core/lib/results/getDisposalTextFromResult/exclusionOrderDisposalText.ts
