+++ packages/core/lib/results/getDisposalTextFromResult/exclusionRequirementsDisposalText.ts
