+++ packages/core/lib/results/getDisposalTextFromResult/index.ts
