+++ packages/core/lib/results/getDisposalTextFromResult/isFailedToAppearWarrantIssued.test.ts
