+++ packages/core/lib/results/getDisposalTextFromResult/isFailedToAppearWarrantIssued.ts
