+++ packages/core/lib/results/getDisposalTextFromResult/licencedPremisesExclusionOrderDisposalText.ts
