+++ packages/core/lib/results/getDisposalTextFromResult/untilFurtherOrderDisposalText.ts
